Made 'setConfig' dynamic.

+++ commands/setConfig.js
@@ -13,42 +13,21 @@ module.exports = {
 		if (!message.member.hasPermission('ADMINISTRATOR'))
 			return message.channel.send(
 				"You don't have the perms to change this! If this needs to be changed then message a server admin."
-			);
-        let inputChange = args.shift().toLowerCase();
+            );
+        let inputHeader = args.shift().toLowerCase();
+        let newHeader = args.join(' ');
+        let serverHeaders = Object.keys(server);
+        let actualHeader = serverHeaders.find((header) => header.toLowerCase() == inputHeader);
+
+        if (!actualHeader)
+            return message.channel.send('Invalid header! `For the list of avaliable headers, take a look at the documentation.\`');
         try {
-            switch (inputChange) {
-                case 'spreadsheetid':
-                    await changeGuildConfig(server, 'spreadsheetId', args.join(' '));
-                    await message.delete();
-                    return message.channel.send(`Successfully set the spreadsheet id!`);
-                case 'rostername': 
-                    await changeGuildConfig(server, 'rosterName', args.join(' '));
-                    return message.channel.send(`Successfully set the roster name to \`${args.join(' ')}\`!`);
-                case 'name':
-                    await changeGuildConfig(server, 'nameHeader', args.join(' '));
-                    return message.channel.send(`Successfully set the name header to \`${args.join(' ')}\`!`);
-                case 'rank':
-                    await changeGuildConfig(server, 'rankHeader', args.join(' '));
-                    return message.channel.send(`Successfully set the rank header to \`${args.join(' ')}\`!`);
-                case 'subdivisionchange':
-                    await changeGuildConfig(server, 'subdivisionChangeHeader', args.join(' '));
-                    return message.channel.send(`Successfully set the subdivision-change header to \`${args.join(' ')}\`!`);
-                case 'lastpromotiondate':
-                    await changeGuildConfig(server, 'lastPromotionDateHeader', args.join(' '));
-                    return message.channel.send(`Successfully set the last promotion-date header to \`${args.join(' ')}\`!`);
-                case 'discord':
-                    await changeGuildConfig(server, 'discordHeader', args.join(' '));
-                    return message.channel.send(`Successfully set the discord header to \`${args.join(' ')}\`!`);
-                case 'status':
-                    await changeGuildConfig(server, 'statusHeader', args.join(' '));
-                    return message.channel.send(`Successfully set the status header to \`${args.join(' ')}\`!`);
-                case 'ranksystem':
-                    await changeGuildConfig(server, 'rankSystem', args.join(' ').toLowerCase());
-                    return message.channel.send(`Successfully set the rank system to \`${args.join(' ').toLowerCase()}\`!`);
-                default:
-                    return message.channel.send(`Invalid arguemnts! Usage: ${server.prefix}${this.name} ${this.usage}`);
-            }
-        } catch (err) {
+            await changeGuildConfig(server, actualHeader, newHeader);
+            if (actualHeader == 'spreadsheetId')
+            await message.delete();
+            return message.channel.send(`Successfully set the \`${actualHeader}\` to \`${actualHeader == 'spreadsheetId' ? 'REDACTED' : newHeader}\`!`);
+        } 
+        catch (err) {
             await sendErrorEmbed(message, { message: `**Command:** ${message.content}\n**Error:** ${err}` });
             return message.channel.send(`Failed to make that change to the config.`);
         }
